[STD-28] Changing injection approach
