+++ src/main/java/com/pine/app/editor/panels/files/FilesTreePanel.java
@@ -1,6 +1,11 @@
 package com.pine.app.editor.panels.files;
 
 import com.pine.app.core.ui.panel.AbstractPanel;
+import com.pine.common.Inject;
+import com.pine.common.fs.FSService;
 
 public class FilesTreePanel  extends AbstractPanel {
+    @Inject
+    private FSService fsService;
+
 }
