+++ src/main/java/com/pine/common/Inject.java
@@ -1,0 +1,8 @@
+package com.pine.common;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Inject {
+}
